requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
flask-restplus==0.13.0
gunicorn==19.10.0
Pillow==6.2.2
werkzeug==0.16.0
